Move evaluator test
